// src/lib/node-multiplexer/nm-common/nm.types.ts
export namespace NodeTypes {
  export interface BlockChainScanContractDetails {
    SourceCode: string;
    ABI: string;
    ContractName: string;
    CompilerVersion: string;
    OptimizationUsed: string;
    Runs: string;
    ConstructorArguments: string;
    EVMVersion: string;
    Library: string;
    LicenseType: string;
    Proxy: string;
    Implementation: string;
    SwarmSource?: string;
  }
}